+++ examples/filter-array-validate-token/index.js
@@ -1,6 +1,6 @@
 import {
     FilterError,
-    HTMLBuildAttributes,
+    Composer,
     createFilterList,
     escapeHTMLEntities,
     filterFallback,
@@ -35,7 +35,7 @@ const filterTokenList = createFilterList((value, name, fallback = false) => {
     return filterFallback(value, name, fallback);
 }, ' ');
 
-const htmlBuildAttributes = new HTMLBuildAttributes(
+const htmlBuildAttributes = new Composer(
     filterTokenList,
     escapeHTMLEntities
 );